+++ ai-feedback-service/src/main/java/com/team9/ai_feedback_service/api/backend/AIQuestionController.java
@@ -1,7 +1,6 @@
 package com.team9.ai_feedback_service.api.backend;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.team9.ai_feedback_service.domain.dto.QuestionResponseDto;
+import com.team9.ai_feedback_service.domain.dto.response.QuestionResponseDto;
 import com.team9.ai_feedback_service.remote.dto.request.QuestionRequestDto;
 import com.team9.ai_feedback_service.service.AIQuestionService;
 import com.team9.common.dto.ApiResponseDto;
