+++ ai-feedback-service/src/main/java/com/team9/ai_feedback_service/domain/dto/response/QuestionResponseDto.java
@@ -1,4 +1,4 @@
-package com.team9.ai_feedback_service.domain.dto;
+package com.team9.ai_feedback_service.domain.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
